Code refactoring Added comments, events are now a separate library
