+++ include/Events.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Events/EventBase.hpp"
+#include "Events/SFMLEvent.hpp"
+#include "Events/TimedEvent.hpp"
+#include "Events/CustomEvent.hpp"
+#include "Events/WindowEventManager.hpp"
