+++ include/Events/CustomEvent.hpp
@@ -1,0 +1,5 @@
+#pragma once
+#include "EventBase.hpp"
+
+template<typename... func_args>
+class CustomEvent : public EventBase<func_args...> {};
