+++ include/Nodes.hpp
@@ -2,6 +2,5 @@
 
 #include "Nodes/Node.hpp"
 #include "Nodes/FPSCounter.hpp"
-#include "Nodes/EventManager.hpp"
 #include "Nodes/RenderManager.hpp"
 #include "Nodes/Application.hpp"
