+++ include/Nodes/ContainerNode.hpp
@@ -3,6 +3,11 @@
 #include "Node.hpp"
 
 template<typename T>
+/*
+    Class template for Nodes with container functionality
+    Children of type other than T can still be added,
+    but wonn't be taken into account when using specific functions
+*/
 class ContainerNode : public Node
 {
     public:
