+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(Nodes)
 add_subdirectory(Utility)
+add_subdirectory(Events)
 
 add_executable(epic_dungeon main.cpp)
 target_include_directories(epic_dungeon PUBLIC ${PROJECT_SOURCE_DIR}/include)
-target_link_libraries(epic_dungeon PUBLIC SFML::Graphics Nodes Utility)
+target_link_libraries(epic_dungeon PUBLIC SFML::Graphics Nodes Events Utility)
